+++ src/Symfony/Component/Mailer/Transport/AbstractTransport.php
@@ -39,6 +39,8 @@ abstract class AbstractTransport implements TransportInterface
         $this->logger = $logger ?: new NullLogger();
     }
 
+    abstract public function getName(): string;
+
     /**
      * Sets the maximum number of messages to send per second (0 to disable).
      */